// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as auth from "../auth.js";
import type * as auth_emailOtp from "../auth/emailOtp.js";
import type * as clients from "../clients.js";
import type * as dashboard from "../dashboard.js";
import type * as expenses from "../expenses.js";
import type * as http from "../http.js";
import type * as inventory from "../inventory.js";
import type * as inventoryCorrections from "../inventoryCorrections.js";
import type * as locations from "../locations.js";
import type * as orders from "../orders.js";
import type * as outlets from "../outlets.js";
import type * as products from "../products.js";
import type * as seedData from "../seedData.js";
import type * as stockMovements from "../stockMovements.js";
import type * as testExpenses from "../testExpenses.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  "auth/emailOtp": typeof auth_emailOtp;
  clients: typeof clients;
  dashboard: typeof dashboard;
  expenses: typeof expenses;
  http: typeof http;
  inventory: typeof inventory;
  inventoryCorrections: typeof inventoryCorrections;
  locations: typeof locations;
  orders: typeof orders;
  outlets: typeof outlets;
  products: typeof products;
  seedData: typeof seedData;
  stockMovements: typeof stockMovements;
  testExpenses: typeof testExpenses;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};